Beta version
This is now the official Beta version.    New stuff:   We have the high level command set and a first pass testing the commands.  The flexpwm library is reorganized and the legacy prototyping routines are removed.  And, we are switching over to our new convention for the command line parsing library.  If you find any issues please direct message or email me.